+++ sdk/containerregistry/azure-containerregistry/tests/asynctestcase.py
@@ -4,13 +4,8 @@
 # Licensed under the MIT License.
 # ------------------------------------
 from azure.containerregistry.aio import (
-    ContainerRepository,
+    # ContainerRepository,
     ContainerRegistryClient,
-)
-from azure.containerregistry import (
-    ArtifactTagProperties,
-    ContentProperties,
-    ArtifactManifestProperties,
 )
 
 from azure.core.credentials import AccessToken