+++ sdk/recoveryservices/azure-mgmt-recoveryservicesbackup/azure/mgmt/recoveryservicesbackup/aio/operations/_backup_crr_job_details_operations.py
@@ -44,7 +44,7 @@ class BackupCrrJobDetailsOperations:
         self,
         azure_region: str,
         parameters: "_models.CrrJobRequest",
-        **kwargs
+        **kwargs: Any
     ) -> "_models.JobResource":
         """Get CRR job details from target region.